rename demo.domains to demo.settings
