+++ demo/settings/abacus/__init__.py
@@ -16,7 +16,7 @@ from proscenium.core import Prop
 from proscenium.verbs.invoke import process_tools
 from proscenium.patterns.tools import apply_tools
 
-import demo.domains.abacus as domain
+import demo.settings.abacus as domain
 from demo.config import default_model_id
 
 logging.getLogger(__name__).addHandler(logging.NullHandler())
