+++ demo/settings/legal/__init__.py
