+++ demo/settings/legal/entity_resolvers.py
